feat!: Standardize configuration options.
Use the CLI entrypoint to parse all arguments rather than looking them
up via the environment in various locations. In addition to improving
consistency, we also get the benefits of generated help documentation
for the CLI.

This is a breaking change because we renamed some environment variables
to avoid exposing Rocket's config structure.