feat: Filter transaction list.
The transaction list can now be filtered by account. It is also
paginated. Since transactions are always listed in the same order, the
pagination is implemented with a cursor pointing to the last seen item.
Providing it to the next request will only return transactions occurring
after the item the cursor refers to.