fix: Remove caching from Docker build.
Our Docker image built the application dependencies first before
building the actual application to utilize the Docker cache in the case
where only application code had changed. With the introduction of our
build script, this was causing problems. Since most of our builds happen
in CI where there is no cache, the simple fix is to just remove the
caching.