ci: Create Sentry release on deploy.
Create a Sentry release when pushing a new version of the API image so
we get better tracking of release creation and timing. If we did not do
this, we would only see releases that cause issues in Sentry, and the
creation times would be the instant the first issue was seen rather than
when the release was pushed.

Closes #139